refactor(kafka): limit kafka metrics by default
